+++ Core-1.8.9/src/cn/yapeteam/yolbi/ui/webui/impl/Handler.java
@@ -1,0 +1,7 @@
+package cn.yapeteam.yolbi.ui.webui.impl;
+
+import net.montoyo.mcef.api.IJSQueryCallback;
+
+public interface Handler {
+    void handle(String[] args, IJSQueryCallback callback);
+}
